Improve build system

+++ build/xref.sh
@@ -9,21 +9,29 @@
 #         containing the script)
 # OUTPUT = Output dir
 # PHPXREF_DIR = phpxref install dir
+#
+# $1 = Config file
+
+if [ -z "$PHPXREF_DIR" ] ; then
+	echo "PHPXREF_DIR is not defined"
+	exit 1
+fi
 
 TMP=/tmp/.t$$
 /bin/rm -rf $TMP
 
-cd `dirname $0`
-dir=`/bin/pwd`
+save_wd=`pwd`
+cd $INPUT
+SOURCE=`pwd`
+cd $save_wd
 
-SOURCE=$INPUT
-echo $SOURCE | grep '^/' >/dev/null 2>&1 || SOURCE=$dir/$INPUT
+(
+echo "SOURCE=$SOURCE"
+echo "OUTPUT=$OUTPUT"
+grep -v '^SOURCE=' <$1 | grep -v 'OUTPUT='
+) >$TMP
 
-echo "SOURCE=$SOURCE" >$TMP
-echo "OUTPUT=$OUTPUT" >>$TMP
-grep -v '^SOURCE=' <$1 | grep -v 'OUTPUT=' >>$TMP
-
-cd $PHPXREF_DIR
+cd $PHPXREF
 perl phpxref.pl -c $TMP
 
 /bin/rm -f $TMP
